[#31] 게시글 작성 두번째 페이지 구현
