+++ Projects/App/Resources/Image/Assets.xcassets/Write.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "new.svg",
+      "filename" : "new (1).svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
