+++ Projects/App/Sources/Feature/HomeFeature/Views/ItemListView.swift
@@ -128,7 +128,7 @@ struct ItemListView: View {
                 }
                 .padding(.horizontal, 24)
 
-                NavigationLink(destination: PostView()) {
+                NavigationLink(destination: PostCreateStep1View()) {
                     Circle()
                         .fill(GwangsanAsset.Color.mainGreen500.swiftUIColor)
                         .frame(width: 60, height: 60)
